+++ src/Pyro4/nsc.py
@@ -78,7 +78,7 @@ def main(args):
     parser.add_option("-n", "--host", dest="host", help="hostname of the NS")
     parser.add_option("-p", "--port", dest="port", type="int",
                       help="port of the NS (or bc-port if host isn't specified)")
-    parser.add_option("-u","--unixsocket", help="unix domain socket name of the NS")
+    parser.add_option("-u","--unixsocket", help="Unix domain socket name of the NS")
     parser.add_option("-v", "--verbose", action="store_true", dest="verbose", help="verbose output")
     options, args = parser.parse_args(args)
     if not args or args[0] not in ("register", "remove", "removematching", "list", "listmatching", "ping"):