+++ src/Pyro4/socketserver/threadpoolserver.py
@@ -208,7 +208,7 @@ class SocketServer_Threadpool(object):
             try:
                 self.sock.close()
                 if type(sockname) is str:
-                    # it was a unix domain socket, remove it from the filesystem
+                    # it was a Unix domain socket, remove it from the filesystem
                     if os.path.exists(sockname):
                         os.remove(sockname)
             except Exception: