+++ src/Pyro4/test/echoserver.py
@@ -70,7 +70,7 @@ def main(args, returnWithoutLooping=False):
     parser=OptionParser()
     parser.add_option("-H","--host", default="localhost", help="hostname to bind server on (default=localhost)")
     parser.add_option("-p","--port", type="int", default=0, help="port to bind server on")
-    parser.add_option("-u","--unixsocket", help="unix domain socket name to bind server on")
+    parser.add_option("-u","--unixsocket", help="Unix domain socket name to bind server on")
     parser.add_option("-n","--naming", action="store_true", default=False, help="register with nameserver")
     parser.add_option("-N","--nameserver", action="store_true", default=False, help="also start a nameserver")
     parser.add_option("-v","--verbose", action="store_true", default=False, help="verbose output")